refactor: extract CJKV full-width abbreviation pattern into base.py mixin
